requirements.txt
alembic==1.18.4
annotated-types==0.7.0
anyio==4.13.0
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
cryptography==46.0.6
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
requests==2.33.0
playwright
pytest-playwright
six==1.17.0
SQLAlchemy==2.0.48
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
volcengine-python-sdk==5.0.17
Werkzeug==3.0.3
tos==2.9.0
